Added showcase scene; Added PBR texture loader; Started working on serialization.

+++ Emerald/src/graphics/buffers/gBuffer.cpp
@@ -6,7 +6,7 @@ GBuffer::GBuffer() : m_fbo(0) {
 
 	m_attributesTexture = m_fbo->AddBuffer("Attributes", TextureParameters(RGBA, RGBA, LINEAR, CLAMP_TO_EDGE, T_UNSIGNED_BYTE));
 	m_colorTexture = m_fbo->AddBuffer("Color", TextureParameters(RGBA, RGBA, LINEAR, CLAMP_TO_EDGE, T_UNSIGNED_BYTE));
-	m_normalTexture = m_fbo->AddBuffer("Normals", TextureParameters(RGBA16, RGBA, LINEAR, CLAMP_TO_EDGE, T_FLOAT));
+	m_normalTexture = m_fbo->AddBuffer("Normals", TextureParameters(RGBA32, RGBA, LINEAR, CLAMP_TO_EDGE, T_FLOAT));
 	m_positionTexture = m_fbo->AddBuffer("Positions", TextureParameters(RGBA32, RGBA, LINEAR, CLAMP_TO_EDGE, T_FLOAT));
 }
